absSum() now in samplingWindow instead of in all image
The calculation of the noise-induced bias produces values that are way too big, resulting in a contrast weighting function with negative value. I suspect that the sum of all absolute pixels value^2 of the Fourier space bandpass filter is done only in the sampling window and not on the whole image. This is what is corrected here.