Finished estimate eta
The cameraOTF, illuminationOTF and detectionOTF returned arrays with only 0s before. I verified the java code for the Fiji plugin and it also returned 0, which is weird, but then in my code I changed the wavelength to mm instead of nm and now it works well. I actually have numbers for the eta function, so this was probably the problem.
Also, for the cameraOTF, I changed the initialization of the np.ndarray to np.float64 instead of np.uint16 and it works now.